Added my examples with the source code used to generate them
Added my sequences.
I wrote these using python so I have added the source python files as well as the baked CSV files. I wrote a version of the neopixel library that can be run without the tree. It visualises the tree using matplotlib and has the option to generate the CSV file from the inputs it was given.
It also adds the baked CSV files for Matt's tree.
Some of these are designed to run at slower frame rates so require https://github.com/standupmaths/xmastree2021/pull/30