Add kuramoto_synchronization.csv
I have created an animation of the tree split into 5 different segments. The state of the LEDs in each segment is dictated by a travelling sine wave. Each segment begins with a sine wave having a unique phase from all other segment waves. As time passes, the phases of the sine waves are adjusted using the Kuramoto model for coupled oscillator synchronization. By the end of the animation, all segments are in phase with each other and appear to be changing as one whole segment. This was really fun to model! I hope you enjoy. If you want to take a closer look at how the animation was acheived, my c# code will be made available in my public repository.